Clean up pass, added utils.py
